fix ios and android issues for build
